Refactor 'chunk' code
