+++ src/lib/services/pdf-generation/view-helpers.js
@@ -5,7 +5,7 @@ const naldDates = require('@envage/water-abstraction-helpers').nald.dates
 
 const naldRegions = require('./data/nald-regions')
 const naldAreas = require('./data/nald-areas')
-const chunk = require('../../chunk.js')
+const { chunk } = require('../../object-helpers.js')
 
 /**
  * Outputs the region name given the NALD region code
