+++ src/modules/batch-notifications/config/returns/lib/return-notification-contacts.js
@@ -4,7 +4,7 @@ const { groupBy } = require('lodash')
 const returnsConnector = require('../../../../../lib/connectors/returns')
 const documentsConnector = require('../../../../../lib/connectors/crm/documents')
 const { createContacts } = require('../../../../../lib/models/factory/crm-contact-list')
-const chunk = require('../../../../../lib/chunk.js')
+const { chunk } = require('../../../../../lib/object-helpers.js')
 
 const groupReturnsByLicenceNumber = returns => groupBy(returns, ret => ret.licence_ref)
